[MYDA-132] Add Connectmydata proto (#27)
* Add AuthService

* Convert connect grpc proto

* Add Exception case

* Add issueTokenClient

* Add comment

* convert constructer to Lombok @RequiredArgsConstructor

* remove external path

* remove blank

* fix bug

* Add approve TestCode

* remove userId

* rename userId to banksaladUserId

+++ oauth/src/main/java/com/banksalad/collectmydata/oauth/dto/UserAuthInfo.java
@@ -8,6 +8,6 @@ import lombok.Getter;
 public class UserAuthInfo {
 
   private String token;
-  private Integer banksaladUserId;
+  private Long banksaladUserId;
   private String os;
 }
